requirements.txt
streamlit
librosa
numpy
soundfile
python-dotenv
pandas
openai
plotly